requirements.txt
django==4.1
python-dotenv
psycopg2-binary
djangorestframework==3.14
djangorestframework-simplejwt
djoser
django-filter